enabled cmake build and tests
