+++ .gitignore
@@ -42,6 +42,7 @@ CMakeCache.txt
 CMakeFiles/
 CMakeLists.txt.user
 build-*/
+build/
 # qtcreator generated files
 *.pro.user*
