+++ CMakeLists.txt
@@ -1,0 +1,10 @@
+cmake_minimum_required (VERSION 3.13)
+#project (algsin81)
+
+set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
+set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
+
+enable_testing()
+add_subdirectory (dev) 
+add_subdirectory (test)
