--- dev/patternmatching/KMP/CMakeLists.txt
@@ -1,4 +1,0 @@
-cmake_minimum_required(VERSION 2.8)
-
-project(KMP)
-add_executable(${PROJECT_NAME} "main.cpp" "kmp.cpp")
